better stream team members
